+++ Pod/CloudyVC.cpp
@@ -32,7 +32,7 @@ void CloudyVC::viewReceiveKeyUp(vector<KeyType> _keyTypes)
 
 void CloudyVC::viewReceiveKeyDown(vector<KeyType> _keyTypes)
 {
-	for (int i = 0; i < _keyTypes.size(); ++i) {
+	for (size_t i = 0; i < _keyTypes.size(); ++i) {
 		if (_keyTypes[i] == KeyType::down
 			&& this->mario->getComponentIdStandingOn() == this->componentIdToThirdMap
 			&& this->mario->getX() >= leftAnchorGreenPipeToPassThrough
@@ -47,10 +47,10 @@ void CloudyVC::viewReceiveKeyDown(vector<KeyType> _keyTypes)
 
 void CloudyVC::viewWillUpdate(float _dt)
 {// Check by cell in grid
-	int beginRow = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight());
-	int endRow = ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight());
-	int beginCol = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth());
-	int endCol = ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth());
+	int beginRow = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()));
+	int endRow = int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()));
+	int beginCol = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()));
+	int endCol = int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()));
 	for (int i = beginRow; i < endRow; ++i) {
 		for (int j = beginCol; j < endCol; ++j) {
 
@@ -98,10 +98,10 @@ void CloudyVC::viewUpdate(float _dt)
 	}
 
 	// Check by cell in grid
-	int beginRow = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight());
-	int endRow = ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight());
-	int beginCol = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth());
-	int endCol = ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth());
+	int beginRow = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()));
+	int endRow = int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()));
+	int beginCol = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()));
+	int endCol = int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()));
 	for (int i = beginRow; i < endRow; ++i) {
 		for (int j = beginCol; j < endCol; ++j) {
 
@@ -267,10 +267,10 @@ void CloudyVC::viewDidUpdate(float _dt)
 	}
 
 	// Check by cell in grid
-	int beginRowMario = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight());
-	int endRowMario = ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight());
-	int beginColMario = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth());
-	int endColMario = ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth());
+	int beginRowMario = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()));
+	int endRowMario = int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()));
+	int beginColMario = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()));
+	int endColMario = int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()));
 	for (int i = beginRowMario; i < endRowMario; ++i) {
 		for (int j = beginColMario; j < endColMario; ++j) {
 
@@ -293,7 +293,7 @@ void CloudyVC::viewDidUpdate(float _dt)
 
 				// Green Pipe
 				else if (beginGreenPipeId <= (*itr)->getId() && (*itr)->getId() <= endGreenPipeId) {
-					this->mario->handleGreenPipeDownCollision(static_cast<GreenPipe*>(*itr), this->componentIdToThirdMap, this->leftAnchorGreenPipeToPassThrough, this->rightAnchorGreenPipeToPassThrough, _dt);
+					this->mario->handleGreenPipeDownCollision(static_cast<GreenPipe*>(*itr), this->componentIdToThirdMap, float(this->leftAnchorGreenPipeToPassThrough), float(this->rightAnchorGreenPipeToPassThrough), _dt);
 				}
 
 				// Gift Brick
@@ -310,18 +310,18 @@ void CloudyVC::viewDidUpdate(float _dt)
 					}
 
 					// Super Mushroom collide to others
-					int beginRowSuperMushroom = floor(((*itr)->getY() - (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight());
-					int endRowSuperMushroom = ceil(((*itr)->getY() + (*itr)->getHeight() + (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight());
-					int beginColSuperMushroom = floor(((*itr)->getX() - (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth());
-					int endColSuperMushroom = ceil(((*itr)->getX() + (*itr)->getWidth() + (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth());
+					int beginRowSuperMushroom = int(floor(((*itr)->getY() - (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight()));
+					int endRowSuperMushroom = int(ceil(((*itr)->getY() + (*itr)->getHeight() + (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight()));
+					int beginColSuperMushroom = int(floor(((*itr)->getX() - (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth()));
+					int endColSuperMushroom = int(ceil(((*itr)->getX() + (*itr)->getWidth() + (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth()));
 
 					beginRowSuperMushroom = beginRowSuperMushroom < 0 ? 0 : beginRowSuperMushroom;
 					endRowSuperMushroom = endRowSuperMushroom > Grid::getInstance()->getTotalRows() ? Grid::getInstance()->getTotalRows() : endRowSuperMushroom;
 					beginColSuperMushroom = beginColSuperMushroom < 0 ? 0 : beginColSuperMushroom;
 					endColSuperMushroom = endColSuperMushroom > Grid::getInstance()->getTotalCols() ? Grid::getInstance()->getTotalCols() : endColSuperMushroom;
 
-					for (int r = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()); r < ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()); ++r) {
-						for (int c = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()); c < ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()); ++c) {
+					for (int r = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight())); r < int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight())); ++r) {
+						for (int c = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth())); c < int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth())); ++c) {
 							unordered_set<Component*> superMushroomCell = Grid::getInstance()->getCell(r, c);
 							unordered_set<Component*> ::iterator superMushroomItr;
 							for (superMushroomItr = superMushroomCell.begin(); superMushroomItr != superMushroomCell.end(); ++superMushroomItr) {
@@ -349,10 +349,10 @@ void CloudyVC::viewDidUpdate(float _dt)
 				// Fire Ball
 				else if (beginFireBallId <= (*itr)->getId() && (*itr)->getId() <= endFireBallId) {
 					// FireBall to others
-					int beginRowFireball = floor(((*itr)->getY() - (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight());
-					int endRowFireball = ceil(((*itr)->getY() + (*itr)->getHeight() + (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight());
-					int beginColFireball = floor(((*itr)->getX() - (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth());
-					int endColFireball = ceil(((*itr)->getX() + (*itr)->getWidth() + (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth());
+					int beginRowFireball = int(floor(((*itr)->getY() - (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight()));
+					int endRowFireball = int(ceil(((*itr)->getY() + (*itr)->getHeight() + (Camera::getInstance()->getHeight() / 2)) / Grid::getInstance()->getCellHeight()));
+					int beginColFireball = int(floor(((*itr)->getX() - (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth()));
+					int endColFireball = int(ceil(((*itr)->getX() + (*itr)->getWidth() + (Camera::getInstance()->getWidth() / 2)) / Grid::getInstance()->getCellWidth()));
 
 					beginRowFireball = beginRowFireball < 0 ? 0 : beginRowFireball;
 					endRowFireball = endRowFireball > Grid::getInstance()->getTotalRows() ? Grid::getInstance()->getTotalRows() : endRowFireball;
@@ -395,8 +395,8 @@ void CloudyVC::viewWillRender()
 			map->Draw(Drawing::getInstance()->getCloudyMapTexture());
 		}
 
-		for (int i = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()); i < ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()); ++i) {
-			for (int j = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()); j < ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()); ++j) {
+		for (int i = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight())); i < int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight())); ++i) {
+			for (int j = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth())); j < int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth())); ++j) {
 				if (Grid::getInstance()->getCell(i, j).size() == 0) continue;
 
 				unordered_set<Component*> cell = Grid::getInstance()->getCell(i, j);
@@ -416,8 +416,8 @@ void CloudyVC::viewWillRender()
 			}
 		}
 
-		for (int i = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()); i < ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()); ++i) {
-			for (int j = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()); j < ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()); ++j) {
+		for (int i = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight())); i < int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight())); ++i) {
+			for (int j = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth())); j < int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth())); ++j) {
 				if (Grid::getInstance()->getCell(i, j).size() == 0) continue;
 
 				unordered_set<Component*> cell = Grid::getInstance()->getCell(i, j);
@@ -435,8 +435,8 @@ void CloudyVC::viewWillRender()
 			mario->Draw();
 		}
 
-		for (int i = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()); i < ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()); ++i) {
-			for (int j = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()); j < ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()); ++j) {
+		for (int i = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight())); i < int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight())); ++i) {
+			for (int j = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth())); j < int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth())); ++j) {
 				if (Grid::getInstance()->getCell(i, j).size() == 0) continue;
 
 				unordered_set<Component*> cell = Grid::getInstance()->getCell(i, j);
@@ -450,8 +450,8 @@ void CloudyVC::viewWillRender()
 			}
 		}
 
-		for (int i = floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight()); i < ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight()); ++i) {
-			for (int j = floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth()); j < ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth()); ++j) {
+		for (int i = int(floor(Camera::getInstance()->getY() / Grid::getInstance()->getCellHeight())); i < int(ceil((Camera::getInstance()->getY() + Camera::getInstance()->getHeight()) / Grid::getInstance()->getCellHeight())); ++i) {
+			for (int j = int(floor(Camera::getInstance()->getX() / Grid::getInstance()->getCellWidth())); j < int(ceil((Camera::getInstance()->getX() + Camera::getInstance()->getWidth()) / Grid::getInstance()->getCellWidth())); ++j) {
 				if (Grid::getInstance()->getCell(i, j).size() == 0) continue;
 
 				unordered_set<Component*> cell = Grid::getInstance()->getCell(i, j);
@@ -490,7 +490,7 @@ void CloudyVC::viewWillRelease()
 void CloudyVC::adaptRangeID(vector<string> data, char seperator)
 {
 	vector<int> v;
-	for (int i = 0; i < data.size(); ++i) {
+	for (size_t i = 0; i < data.size(); ++i) {
 		if (i == 0) {
 			v = Tool::splitToVectorIntegerFrom(data[i], seperator);
 			this->beginGroundId = v[0];
@@ -575,7 +575,7 @@ void CloudyVC::adaptData()
 			continue;
 		}
 		else if (line == "</Grounds>") {
-			for (int i = 0; i < data.size(); ++i) {
+			for (size_t i = 0; i < data.size(); ++i) {
 				Ground* ground = new Ground(0, 0, 0, 0, 0, 0, 0, 0);
 				ground->load(data[i], ',');
 				this->grounds->push_back(ground);
@@ -587,7 +587,7 @@ void CloudyVC::adaptData()
 			continue;
 		}
 		else if (line == "</CoinFrames>") {
-			for (int i = 0; i < data.size(); ++i) {
+			for (size_t i = 0; i < data.size(); ++i) {
 				Coin* coin = new Coin(0, 0, 0, 0, 0, 0, 0, 0);
 				coin->loadInfo(data[i], ',');
 				this->coins->insert(coin);
@@ -599,7 +599,7 @@ void CloudyVC::adaptData()
 			continue;
 		}
 		else if (line == "</GreenPipeFrames>") {
-			for (int i = 0; i < data.size(); ++i) {
+			for (size_t i = 0; i < data.size(); ++i) {
 				GreenPipe* greenPipe = new GreenPipe(0, 0, 0, 0, 0, 0, 0);
 				greenPipe->loadInfo(data[i], ',');
 				greenPipes->push_back(greenPipe);
@@ -611,7 +611,7 @@ void CloudyVC::adaptData()
 			continue;
 		}
 		else if (line == "</GiftBrickFrames>") {
-			for (int i = 0; i < data.size(); ++i) {
+			for (size_t i = 0; i < data.size(); ++i) {
 				GiftBrick* giftBrick = new GiftBrick(0, 0, 0, 0, 0, 0);
 				giftBrick->loadInfo(data[i], ',');
 				giftBricks->push_back(giftBrick);
@@ -670,7 +670,7 @@ void CloudyVC::adaptAnimation()
 	}
 
 	// Green Pipes
-	for (int i = 0; i < this->greenPipes->size(); ++i) {
+	for (size_t i = 0; i < this->greenPipes->size(); ++i) {
 		if (this->greenPipes->at(i)->getFloorNumber() == 2) {
 			this->greenPipes->at(i)->setAnimation(new Animation(AnimationBundle::getInstance()->getGreenPipe2Floor()));
 		}
@@ -680,7 +680,7 @@ void CloudyVC::adaptAnimation()
 	}
 
 	// Gift Bricks
-	for (int i = 0; i < this->giftBricks->size(); ++i) {
+	for (size_t i = 0; i < this->giftBricks->size(); ++i) {
 		this->giftBricks->at(i)->setState(GiftBrickState::FULLGIFTBRICK);
 		if (this->giftBricks->at(i)->getGiftType() == SuperMushroomOrSuperLeaf) { // Super Mushroom, super leaf
 			this->giftBricks->at(i)->getSuperMushroom()->setUpAnimation();
@@ -704,7 +704,7 @@ void CloudyVC::adaptAnimation()
 void CloudyVC::adaptToGrid()
 {
 	// Grounds
-	for (int i = 0; i < this->grounds->size(); ++i) {
+	for (size_t i = 0; i < this->grounds->size(); ++i) {
 		Grid::getInstance()->add(this->grounds->at(i));
 	}
 
@@ -715,12 +715,12 @@ void CloudyVC::adaptToGrid()
 	}
 
 	// Green Pipes
-	for (int i = 0; i < this->greenPipes->size(); ++i) {
+	for (size_t i = 0; i < this->greenPipes->size(); ++i) {
 		Grid::getInstance()->add(this->greenPipes->at(i));
 	}
 
 	// Gift Bricks
-	for (int i = 0; i < this->giftBricks->size(); ++i) {
+	for (size_t i = 0; i < this->giftBricks->size(); ++i) {
 		Grid::getInstance()->add(this->giftBricks->at(i));
 	}
 }
